Implement TOKEN_DOT_DOT and TOKEN_STEP
I did't like how we previously handled iteration in nue because doing an iteration like 1 to 100 for example relied on the presence of a predefined function, range(). Not only does this serve as a security issue because a malicious actor could overwrite the range() function (or an unsuspecting user could break core language functionality), and it seems absurd that such an important language feature needs an entirely separate function rather than having methods for iteration be integrated in the syntax itself.