Removal of old TODOs
Removed a few TODO comments which were no longer relevant

Also removed this, dont get confused:
// TODO: get variables names and put them into the lexeme or literal property

the reason this comment was removed is because I somehow forgot that we already store the literal value of a lexeme in a token, which essentially already is the "name of the variable". It will be used by the AST and maybe other later stages.