Update vscode settings
- Finally removed settings relating to swift in vscode.
Swift was only "used" during early development because when Nue code was opened in VSCode, it was automatically detected as Swift. Additionally, the syntax was close enough and highlighting was alright. However, as Nue, its features and syntax evolved, Swift no longer seemed similar and an extension was bound to created for Nue either way.
- Added .vscode/extensions.json which recommends developers to install the official Nue VScode extension when working, developing and testing Nue